sign_language_app/requirements.txt
torch==2.8.0
torchvision==0.19.0
torchaudio==2.8.0

opencv-python==4.12.0.88
mediapipe==0.10.14

streamlit==1.39.0
streamlit-webrtc==0.47.9

numpy==2.2.6
pillow==11.3.0
protobuf==4.25.8


